CODERUB: Separate Logic Test For RetrieveAll Operation - StudentRegistrations
