+++ OtripleS.Web.Api.Tests.Unit/Services/Foundations/StudentRegistrations/StudentRegistrationServiceTests.Logic.cs
@@ -14,40 +14,7 @@ using Xunit;
 namespace OtripleS.Web.Api.Tests.Unit.Services.Foundations.StudentRegistrations
 {
     public partial class StudentRegistrationServiceTests
-    {
-        [Fact]
-        public void ShouldRetrieveAllStudentRegistrations()
-        {
-            // given
-            DateTimeOffset randomDateTime = GetRandomDateTime();
-            IQueryable<StudentRegistration> randomStudentRegistrations =
-                CreateRandomStudentRegistrations(randomDateTime);
-
-            IQueryable<StudentRegistration> storageStudentRegistrations =
-                randomStudentRegistrations;
-
-            IQueryable<StudentRegistration> expectedStudentRegistrations =
-                storageStudentRegistrations;
-
-            this.storageBrokerMock.Setup(broker =>
-                broker.SelectAllStudentRegistrations())
-                    .Returns(storageStudentRegistrations);
-
-            // when
-            IQueryable<StudentRegistration> actualStudentRegistrations =
-                this.studentRegistrationService.RetrieveAllStudentRegistrations();
-
-            // then
-            actualStudentRegistrations.Should().BeEquivalentTo(expectedStudentRegistrations);
-
-            this.storageBrokerMock.Verify(broker =>
-                broker.SelectAllStudentRegistrations(),
-                    Times.Once);
-
-            this.storageBrokerMock.VerifyNoOtherCalls();
-            this.loggingBrokerMock.VerifyNoOtherCalls();
-        }
-
+    {       
         [Fact]
         public async Task ShouldAddStudentRegistrationAsync()
         {
